requirements.txt
fastapi==0.115.2
uvicorn==0.31.1
pydantic==2.9.2  
sqlalchemy==1.4.54
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pytest==8.3.3
pytest-cov==6.0.0
pytest-asyncio==0.24.0
httpx==0.27.0
